+++ test/test_asyncredis.py
@@ -58,7 +58,7 @@ async def test_transaction_fail(async_redis: redis.asyncio.Redis):
             await tr.execute()
 
 
-async def test_pubsub(async_redis, event_loop):
+async def test_pubsub(async_redis):
     queue = asyncio.Queue()
 
     async def reader(ps):
@@ -71,7 +71,7 @@ async def test_pubsub(async_redis, event_loop):
 
     async with async_timeout(5), async_redis.pubsub() as ps:
         await ps.subscribe("channel")
-        task = event_loop.create_task(reader(ps))
+        task = asyncio.get_running_loop().create_task(reader(ps))
         await async_redis.publish("channel", "message1")
         await async_redis.publish("channel", "message2")
         result1 = await queue.get()
@@ -117,14 +117,14 @@ async def test_blocking_timeout(conn):
 
 
 @pytest.mark.slow
-async def test_blocking_unblock(async_redis, conn, event_loop):
+async def test_blocking_unblock(async_redis, conn):
     """Blocking command that gets unblocked after some time."""
 
     async def unblock():
         await asyncio.sleep(0.1)
         await async_redis.rpush("list", "y")
 
-    task = event_loop.create_task(unblock())
+    task = asyncio.get_running_loop().create_task(unblock())
     result = await conn.blpop("list", timeout=1)
     assert result == (b"list", b"y")
     await task